requirements.txt
# Dependencies
fastapi
uvicorn
transformers
torch
openai
pydantic
python-dotenv
google-generativeai
